requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.40
chromadb>=0.4.0
sentence-transformers>=2.2.0
jira>=3.6.0
python-dotenv>=1.0.0
